# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "NimeStream API documentation",
        "contact": {},
        "version": "1.0.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/movie/episodes": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movie"
                ],
                "summary": "Create a new movie episodes",
                "parameters": [
                    {
                        "description": "Request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_muhammadsaefulr_NimeStreamAPI_internal_domain_dto_movie_episode_request.CreateMovieEpisodes"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_muhammadsaefulr_NimeStreamAPI_internal_domain_dto_util_response.SuccessWithDetail-github_com_muhammadsaefulr_NimeStreamAPI_internal_domain_model_MovieEpisode"
                        }
                    }
                }
            }
        },
...
        "/movie/episodes/lists/{movie_id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movie"
                ],
                "summary": "Get all movie episodes by movie ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Movie ID",
                        "name": "movie_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_muhammadsaefulr_NimeStreamAPI_internal_domain_dto_util_response.SuccessWithPaginate-github_com_muhammadsaefulr_NimeStreamAPI_internal_domain_model_MovieEpisode"
                        }
                    }
                }
            }
        },
        "/movie/episodes/upload": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movie"
                ],
                "summary": "Upload new movie episode (with file)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Episode ID",
                        "name": "movie_eps_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Movie ID",
                        "name": "movie_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Title",
                        "name": "title",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Resolution",
                        "name": "resolution",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Video File",
                        "name": "file_video",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_muhammadsaefulr_NimeStreamAPI_internal_domain_dto_util_response.SuccessWithDetail-github_com_muhammadsaefulr_NimeStreamAPI_internal_domain_model_MovieEpisode"
                        }
                    }
                }
            }
        },
        "/movie/episodes/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movie"
                ],
                "summary": "Update a movie episodes",
                "parameters": [
                    {
                        "type": "string",
                        "description": "MovieEps ID (string)",
                        "name": "movie_eps_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_muhammadsaefulr_NimeStreamAPI_internal_domain_dto_movie_episode_request.UpdateMovieEpisodes"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/movie/episodes/{movie_eps_id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movie"
                ],
                "summary": "Get a movie episodes by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "MovieEps ID (string)",
                        "name": "movie_eps_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_muhammadsaefulr_NimeStreamAPI_internal_domain_dto_util_response.SuccessWithDetail-github_com_muhammadsaefulr_NimeStreamAPI_internal_domain_model_MovieEpisode"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movie"
                ],
                "summary": "Delete a movie episodes",
                "parameters": [
                    {
                        "type": "string",
                        "description": "MovieEps ID (string)",
                        "name": "movie_eps_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/movie/episodes/{movie_id}/{movie_eps_id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movie"
                ],
                "summary": "Get all movie episodes by movie ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Movie ID",
                        "name": "movie_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Movie Eps ID",
                        "name": "movie_eps_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_muhammadsaefulr_NimeStreamAPI_internal_domain_dto_util_response.SuccessWithDetail-github_com_muhammadsaefulr_NimeStreamAPI_internal_domain_dto_movie_episode_response_MovieEpisodeResponses"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_muhammadsaefulr_NimeStreamAPI_internal_domain_dto_movie_episode_request.CreateMovieEpisodes": {
            "type": "object",
            "properties": {
                "movie_eps_id": {
                    "type": "string",
                    "example": "wu-nao-monu-eps-1"
                },
                "movie_id": {
                    "type": "string",
                    "example": "wu-nao-monu"
                },
                "resolution": {
                    "type": "string",
                    "example": "720p"
                },
                "title": {
                    "type": "string",
                    "example": "upload"
                },
                "video_url": {
                    "type": "string",
                    "example": "youtube"
                }
            }
        },
        "github_com_muhammadsaefulr_NimeStreamAPI_internal_domain_dto_movie_episode_request.UpdateMovieEpisodes": {
            "type": "object",
            "properties": {
                "movie_id": {
                    "type": "string"
                },
                "resolution": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "github_com_muhammadsaefulr_NimeStreamAPI_internal_domain_dto_movie_episode_response.MovieEpisodeResponses": {
            "type": "object",
            "properties": {
                "current_ep": {
                    "type": "string"
                },
                "detail_url": {
                    "type": "string"
                },
                "episodes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_muhammadsaefulr_NimeStreamAPI_internal_domain_dto_movie_episode_response.SourcesData"
                    }
                },
                "next_ep_url": {
                    "type": "string"
                },
                "release_date": {
                    "type": "string"
                },
                "sources": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_muhammadsaefulr_NimeStreamAPI_internal_domain_dto_movie_episode_response.Sources"
                    }
                },
                "thumbnail_url": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "github_com_muhammadsaefulr_NimeStreamAPI_internal_domain_dto_movie_episode_response.Sources": {
            "type": "object",
            "properties": {
                "data_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_muhammadsaefulr_NimeStreamAPI_internal_domain_dto_movie_episode_response.SourcesData"
                    }
                },
                "movie_eps_id": {
                    "type": "string"
                },
                "res": {
                    "type": "string"
                }
            }
        },
        "github_com_muhammadsaefulr_NimeStreamAPI_internal_domain_dto_movie_episode_response.SourcesData": {
            "type": "object",
            "properties": {
                "title": {
                    "type": "string"
                },
                "video_url": {
                    "type": "string"
                }
            }
        },
...
        "github_com_muhammadsaefulr_NimeStreamAPI_internal_domain_dto_util_response.SuccessWithDetail-github_com_muhammadsaefulr_NimeStreamAPI_internal_domain_dto_movie_episode_response_MovieEpisodeResponses": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/github_com_muhammadsaefulr_NimeStreamAPI_internal_domain_dto_movie_episode_response.MovieEpisodeResponses"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "github_com_muhammadsaefulr_NimeStreamAPI_internal_domain_dto_util_response.SuccessWithDetail-github_com_muhammadsaefulr_NimeStreamAPI_internal_domain_model_MovieEpisode": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/github_com_muhammadsaefulr_NimeStreamAPI_internal_domain_model.MovieEpisode"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "github_com_muhammadsaefulr_NimeStreamAPI_internal_domain_dto_util_response.SuccessWithPaginate-github_com_muhammadsaefulr_NimeStreamAPI_internal_domain_model_MovieEpisode": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_muhammadsaefulr_NimeStreamAPI_internal_domain_model.MovieEpisode"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "page": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "total_pages": {
                    "type": "integer"
                },
                "total_results": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_muhammadsaefulr_NimeStreamAPI_internal_domain_model.MovieEpisode": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "movie_eps_id": {
                    "type": "string"
                },
                "movie_id": {
                    "type": "string"
                },
                "resolution": {
                    "type": "string"
                },
                "source_by": {
                    "type": "string"
                },
                "title": {
                    "description": "title untuk penentu source nya misal pixeldrain atau uploads",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "video_url": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Example Value: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}